Rename all models to lower case
